+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
+## [Unreleased]
+
+### Fixed
+
+- models: Fixed `CachedModel` preloading.
+
 ## [8.0.0b3] - 2024-07-04
 
 ### Added